refactor: move types to their own files
