+++ src/enums.ts
@@ -1,0 +1,46 @@
+export enum LoginProvider {
+  Google,
+  EmailOTP,
+  EmailPassword,
+}
+
+export enum Size {
+  None = "None",
+  Tall = "Tall",
+  Grande = "Grande",
+  Venti = "Venti",
+}
+
+export enum Milk {
+  None = "None",
+  Cow_s = "Cow's",
+  Lactose_Free = "Lactose-Free",
+  Skimmed = "Skimmed",
+  Vegetable = "Vegetable",
+}
+
+export enum Syrup {
+  None = "None",
+  Amaretto = "Amaretto",
+  Caramel = "Caramel",
+  Hazelnut = "Hazelnut",
+  Irish_Cream = "Irish Cream",
+  Vanilla = "Vanilla",
+}
+
+export enum Additive {
+  Ceylon_Cinnamon = "Ceylon Cinnamon",
+  Grated_Chocolate = "Grated Chocolate",
+  Liquid_Chocolate = "Liquid Chocolate",
+  Marshmallow = "Marshmallow",
+  Whipped_Cream = "Whipped Cream",
+  Nutmeg = "Nutmeg",
+  Ice_Cream = "Ice Cream",
+}
+
+export enum Ice {
+  None = "None",
+  Light = "Light",
+  Normal = "Normal",
+  Extra = "Extra",
+}
