+++ src/pages/Orders/HistoryOrders.tsx
@@ -12,8 +12,7 @@ import {
   useCollectionData,
   useDocument,
 } from "react-firebase-hooks/firestore";
-
-import { DeliveryStatusType } from "../../types";
+import { DeliveryStatusType } from "@/types/delivery.types";
 import Empty from "../../components/Empty";
 import { OrderConvert } from "../../converters/orders";
 import OrderItem from "./OrderItem";
