+++ src/types/avatar.types.ts
@@ -1,0 +1,12 @@
+export enum SystemAvatarNames {
+  Man = "Man",
+  ManAlt = "ManAlt",
+  Woman = "Woman",
+  WomanAlt = "WomanAlt"
+}
+
+export interface Avatar {
+  name: SystemAvatarNames | string,
+  path: string,
+  system: boolean,
+}
